[거래] 채팅 거래 관련 예외처리, 거래내역 노출조건 수정

+++ src/main/java/com/zeromarket/server/api/dto/order/TradeHistoryResponse.java
@@ -18,7 +18,7 @@ public class TradeHistoryResponse {
     private String productTitle;
     private String thumbnailUrl;
 
-    private Integer sellPrice;
+    private Long sellPrice;
 
     private TradeType tradeType;
     private TradeStatus tradeStatus;
